don't build --with-async-mode w/ libusb-compat-0.1
Hello,

attached is a little patch that prohibits libftdi from compiling against
libusb-compat-0.1 (the wrapper that provides a libusb-0.1 compatible api
for systems with libusb-1.0) when libftdi is configured with async mode.

unfortunately fixing async mode for libusb-1.0+libusb-compat-0.1 is not
trivial because libusb-1.0 comes with an os abstraction layer that hides
the unix filedescriptor much deeper in internal structures. but failing to
build is imo much better than segfaulting at runtime.

yours,
 - clifford

PS: The USB_CLASS_PTP define is not present in any libusb-0.1 release
and has been added to libusb-compat-0.1 only afaics.