libftdi: add a --without-examples configure flag
There's a few places where I'm only concerned with building the libftdi
library (for integrating it into a JTAG program), so the examples are just
noise to me.  As such, add an --without-examples configure flag so people
can easily disable the subdir.  The default is unchanged -- examples are
still compiled & installed.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>